+++ st2actions/st2actions/runners/remote_command_runner.py
@@ -108,6 +108,7 @@ class ParamikoRemoteCommandRunner(BaseParallelSSHRunner):
                                            user=self._username,
                                            password=self._password,
                                            private_key=self._private_key,
+                                           passphrase=self._passphrase,
                                            hosts=self._hosts,
                                            parallel=self._parallel,
                                            sudo=self._sudo,