+++ CHANGELOG.rst
@@ -4,6 +4,7 @@ Changelog
 in development
 --------------
 
+* Passphrase support for the SSH runner. (improvement)
 * Improvements to ChatOps deployments of packs via ``pack deploy`` [Jon Middleton]
 * Add ``extra`` field to the ActionAlias schema for adapter-specific parameters. (improvement)
 * Dev environment by default now uses gunicorn to spin API and AUTH processes. (improvement)